new file:  Jupyter Notebook
Notebook to explore the different pre-trained word embeddings available.  This notebook currently looks at the 50-dimensional vectors from GloVe.  As a test, the first 50 words from the data set are loaded into Jupyter.  A list of available words is created, a list of the vectors are created, and a dictionary to go between word & vector index is created.